requirements.txt
Pillow
flake8
matplotlib
numpy
opencv-python
pyinstaller
pytesseract
pytest